Extract function for serializing a (key, value) pair to a row of matrix and vice versa into a new module
