+++ src/lib.rs
@@ -1,2 +1,3 @@
 mod binary_fuse_filter;
 mod matrix;
+mod serialization;
