+++ src/matrix.rs
@@ -1,4 +1,4 @@
-use crate::binary_fuse_filter::{self, BinaryFuseFilter};
+use crate::{binary_fuse_filter, serialization};
 use sha3::{
     digest::{ExtendableOutput, Update, XofReader},
     Shake128,
@@ -90,7 +90,7 @@ impl Matrix {
                     let key = *hash_to_key.get(&hash)?;
                     let value = *db.get(key)?;
 
-                    let (h0, h1, h2) = BinaryFuseFilter::hash_batch(hash, filter.segment_length, filter.segment_count_length);
+                    let (h0, h1, h2) = binary_fuse_filter::BinaryFuseFilter::hash_batch(hash, filter.segment_length, filter.segment_count_length);
 
                     let found = reverse_h[i] as usize;
                     h012[0] = h0;
@@ -99,7 +99,7 @@ impl Matrix {
                     h012[3] = h012[0];
                     h012[4] = h012[1];
 
-                    let row = binary_fuse_filter::encode_kv_as_row(key, value, mat_elem_bit_len, cols);
+                    let row = serialization::encode_kv_as_row(key, value, mat_elem_bit_len, cols);
 
                     let mat_row_idx0 = h012[found + 0] as usize;
                     let mat_row_idx1 = h012[found + 1] as usize;
@@ -115,7 +115,7 @@ impl Matrix {
                             (elem_idx, elem.wrapping_sub(f2) & mat_elem_mask)
                         })
                         .map(|(elem_idx, elem)| {
-                            let mask = (BinaryFuseFilter::mix(hash, elem_idx as u64) as u32) & mat_elem_mask;
+                            let mask = (binary_fuse_filter::BinaryFuseFilter::mix(hash, elem_idx as u64) as u32) & mat_elem_mask;
                             elem.wrapping_sub(mask) & mat_elem_mask
                         })
                         .collect::<Vec<u32>>();
